♻_refactor(OAuth): grouped db functionality
